requirements.txt
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
pre-commit==3.3.3
psycopg==3.1.9
celery==5.3.0
redis==4.5.5
pdfkit==1.0.0
requests==2.31.0
PyPDF2==3.0.1
boto3==1.26.161
django-storages==1.13.2
django-dynamic-preferences==1.15.0
drf-yasg==1.21.7
